+++ src/onegov.town/onegov/town/views/resource.py
@@ -196,7 +196,7 @@ def handle_cleanup_allocations(self, request, form):
 def get_reservations(self, request):
     return [
         utils.ReservationInfo(reservation, request).as_dict() for reservation
-        in self.bound_reservations(request)
+        in self.bound_reservations(request).all()
     ]